pyproject.toml
[project]
name = "pdf-extractor"
version = "0.1.0"
description = "Expense tracking pipeline — PDF/image extraction to BigQuery via ADK + Gemini."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-adk>=1.27.1",
    "google-genai>=1.67.0",
    "python-dotenv>=1.2.2",
    "google-cloud-bigquery>=3.0.0",
    "google-cloud-storage>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (require real GCP credentials)",
]
testpaths = ["tests"]
